requirements.txt
psycopg2-binary
weasyprint
openai
tiktoken
markdown
beautifulsoup4
